job.goを分解&internalへ移動 #155
