+++ internal/new-video/main_test.go
@@ -1,8 +1,12 @@
-package nsa
+package newvideo
 
 import (
+	"fmt"
+	"os"
 	"testing"
 
+	"github.com/aopontann/niji-tuu/internal/common/db"
+	"github.com/aopontann/niji-tuu/internal/common/youtube"
 	"github.com/joho/godotenv"
 )
 
@@ -25,21 +29,21 @@ func TestNewVideoWebHook(t *testing.T) {
 	}
 }
 
-func TestCreateTaskToNoficationByDiscord(t *testing.T) {
+func TestGetStatusChengedVtubers(t *testing.T) {
 	godotenv.Load(".env.test")
-
-	vids := []string{"EgaXyUcsM48"}
-	err := CreateTaskToNoficationByDiscord(vids)
+	yt, err := youtube.NewYoutube(os.Getenv("YOUTUBE_API_KEY"))
 	if err != nil {
 		t.Error(err)
 	}
-}
-
-func TestDiscordAnnounceJob(t *testing.T) {
-	godotenv.Load(".env")
-	// 新しく動画をアップロードしたプレイリスト情報を取得
-	err := DiscordAnnounceJob("cOaucoqw1Rs")
+	db, err := db.NewDB(os.Getenv("DSN"))
 	if err != nil {
 		t.Error(err)
 	}
+	defer db.Close()
+
+	vtuber, err := GetStatusChengedVtubers(yt, db)
+	if err != nil {
+		t.Error(err)
+	}
+	fmt.Println(vtuber)
 }
